+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Default value for .netrc file path shown in help
 
 ## [0.1.0] - 2021-05-27
 Initial release